[Add] 프론트엔드 폴더 구조 생성
전체적인 폴더 페이지 구성 완료
