+++ src/App.js
@@ -1,21 +1,5 @@
-import "./App.css";
-
-import { Button } from "@mui/material";
-import { signout } from "./service/UserService";
-
 function App() {
-  return (
-    <div className="App">
-      <Button
-        type="submit"
-        variant="contained"
-        color="primary"
-        onClick={signout}
-      >
-        로그아웃
-      </Button>
-    </div>
-  );
+    return <div className="App"></div>;
 }
 
 export default App;
